+++ src/invoke_training/config/pipelines/finetune_lora_config.py
@@ -6,7 +6,6 @@ from invoke_training.config.pipelines.base_pipeline_config import BasePipelineCo
 from invoke_training.config.shared.data.data_loader_config import (
     DreamboothSDDataLoaderConfig,
     ImageCaptionSDDataLoaderConfig,
-    ImagePairPreferenceSDDataLoaderConfig,
 )
 from invoke_training.config.shared.optimizer.optimizer_config import OptimizerConfig
 
@@ -21,18 +20,6 @@ class LoRATrainingConfig(BasePipelineConfig):
 
     hf_variant: str | None = "fp16"
     """The Hugging Face Hub model variant to use. Only applies if `model` is a Hugging Face Hub model name.
-    """
-
-    initial_lora: str | None = None
-    """The LoRA checkpoint directory to initialize the LoRA weights from.
-
-    If set, the following configuration parameters are ignored:
-    - `train_unet`: The UNet will be trained if it is present in `initial_lora`.
-    - `train_text_encoder`: The text encoder will be trained if it is present in `initial_lora`.
-    - `lora_rank_dim`: The LoRA rank dimension from `initial_lora` will be used.
-
-    Currently only LoRA checkpoints in the internal `invoke-training` PEFT format are supported (i.e. checkpoints
-    generated by an `invoke-training` training pipeline).
     """
 
     train_unet: bool = True
@@ -168,17 +155,4 @@ class FinetuneLoRASDXLConfig(LoRATrainingConfig):
     """The name of the Hugging Face Hub VAE model to train against. This will override the VAE bundled with the base
     model (specified by the `model` parameter). This config option is provided for SDXL models, because SDXL shipped
     with a VAE that produces NaNs in fp16 mode, so it is common to replace this VAE with a fixed version.
-    """
-
-
-class DirectPreferenceOptimizationLoRASDConfig(LoRATrainingConfig):
-    type: Literal["DIRECT_PREFERENCE_OPTIMIZATION_LORA_SD"] = "DIRECT_PREFERENCE_OPTIMIZATION_LORA_SD"
-    optimizer: OptimizerConfig
-    data_loader: ImagePairPreferenceSDDataLoaderConfig
-
-    beta: float = 5000.0
-    """The beta parameter, as defined in (https://arxiv.org/pdf/2311.12908.pdf). Larger beta values increase the
-    KL-Divergence penalty, discouraging divergence from the reference model weights.
-
-    Typical values for `beta` are in the range [1000.0, 10000.0].
     """
