+++ src/invoke_training/config/pipelines/pipeline_config.py
@@ -2,8 +2,8 @@ from typing import Annotated, Union
 
 from pydantic import Field
 
+from invoke_training.config._experimental.dpo.config import DirectPreferenceOptimizationLoRASDConfig
 from invoke_training.config.pipelines.finetune_lora_config import (
-    DirectPreferenceOptimizationLoRASDConfig,
     FinetuneLoRASDConfig,
     FinetuneLoRASDXLConfig,
 )
